src/requirements.txt
accelerate==0.29.3
aiohttp==3.9.3
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anonLLM==0.1.10
anyio==3.7.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blis==0.7.11
build==1.1.1
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
coloredlogs==15.0.1
confection==0.1.5
crontab==0.23.0
cryptography==3.4.7
cymem==2.0.8
dataclasses-json==0.6.4
datasets==2.19.0
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Faker==25.3.0
fastapi==0.100.1
filelock==3.13.1
fire==0.4.0
flatbuffers==24.3.7
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
gym-notices==0.0.8
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.3.2
kubernetes==29.0.0
langcodes==3.4.0
language_data==1.2.0
llama_cpp_python==0.2.7
llama-index==0.10.31
llama-index-agent-openai==0.2.2
llama-index-cli==0.1.11
llama-index-core==0.10.31
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.14
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.15
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.0
llamaindex-py-client==0.1.18
Mako==1.3.5
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
neo4j==5.19.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.0
onnxruntime==1.16.3
openai==1.16.2
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==24.0
pandas==2.2.1
peft==0.10.0
pillow==10.3.0
pip==23.3.1
platformdirs==4.2.2
posthog==3.5.0
preshed==3.0.9
protobuf==4.25.3
psutil==5.9.8
pulsar-client==3.4.0
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.17.2
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
PyMySQL==1.0.3
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.0.0
python-crontab==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==4.2.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
sentencepiece==0.2.0
sentry-sdk==2.11.0
setproctitle==1.3.3
setuptools==65.5.1
shtab==1.7.1
simplejson==3.17.6
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.49
SQLAlchemy-Utils==0.41.2
srsly==2.4.8
sseclient==0.0.27
starlette==0.27.0
striprtf==0.0.26
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
thinc==8.2.3
tiktoken==0.6.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.2.2
tqdm==4.66.2
transformers==4.43.3
trl==0.8.6
typer==0.9.4
typing_extensions==4.10.0
typing-inspect==0.9.0
tyro==0.8.3
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.28.0
uvloop==0.19.0
wandb==0.17.5
wasabi==1.1.3
watchfiles==0.21.0
weasel==0.3.4
websocket-client==1.7.0
websockets==12.0
wheel==0.41.2
wikipedia==1.4.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
opencv-python
alfworld
